stack: Fix temporary deadlock in disconnect logic
The Stack.disconnect method calls the IPConnection.disconnect method
with a timeout. The IPConnection.disconnect method synchronously triggers
a disconnected callback that is registered to the Stack.handleConnect
method. The Stack.handleConnect method in turn calls the Stack.disconnect
method again, that is synchronized. This now deadlocks until the timeout
in the first call to the Stack.disconnect method expires.

Fix this by not calling the Stack.disconnect method from the
Stack.handleConnect method. Instead just trigger the desired side effect
of clearing the sensor list.