Shutdown ExecutorService to avoid keeping user-threads around too long
The TimeoutExecutor creates a new ExecutorService per execute call. The
ExecutorService seems to use user-threads and keeps them around for a
moment after executing a task to reuse them for the next task that might
come around.

This has the disadvantage of keeping the Java process running for a moment
even after the main method has already exited. Avoid this by shutting down
the ExecutorService after the TimeoutExecutor is done with it.